+++ driver-core/src/main/com/mongodb/bulk/BulkWriteError.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.client;
+package com.mongodb.bulk;
 
 import org.bson.BsonDocument;
