+++ driver-core/src/main/com/mongodb/bulk/BulkWriteResult.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.client.result;
+package com.mongodb.bulk;
 
 import java.util.List;
