+++ driver-core/src/main/com/mongodb/operation/BaseWriteOperation.java
@@ -25,9 +25,9 @@ import com.mongodb.async.SingleResultCallback;
 import com.mongodb.async.SingleResultFuture;
 import com.mongodb.binding.AsyncWriteBinding;
 import com.mongodb.binding.WriteBinding;
-import com.mongodb.client.BulkWriteError;
-import com.mongodb.client.BulkWriteException;
-import com.mongodb.client.result.BulkWriteResult;
+import com.mongodb.bulk.BulkWriteError;
+import com.mongodb.bulk.BulkWriteException;
+import com.mongodb.bulk.BulkWriteResult;
 import com.mongodb.connection.Connection;
 import com.mongodb.protocol.AcknowledgedWriteResult;
 import com.mongodb.protocol.WriteCommandProtocol;
