+++ driver-core/src/main/com/mongodb/operation/DeleteOperation.java
@@ -18,7 +18,7 @@ package com.mongodb.operation;
 
 import com.mongodb.MongoNamespace;
 import com.mongodb.WriteConcern;
-import com.mongodb.client.result.BulkWriteResult;
+import com.mongodb.bulk.BulkWriteResult;
 import com.mongodb.protocol.DeleteCommandProtocol;
 import com.mongodb.protocol.DeleteProtocol;
 import com.mongodb.protocol.WriteCommandProtocol;
