+++ driver-core/src/test/unit/com/mongodb/protocol/BulkWriteBatchCombinerSpecification.groovy
@@ -17,10 +17,10 @@
 package com.mongodb.protocol
 
 import com.mongodb.ServerAddress
-import com.mongodb.client.BulkWriteError
-import com.mongodb.client.BulkWriteException
+import com.mongodb.bulk.BulkWriteError
+import com.mongodb.bulk.BulkWriteException
+import com.mongodb.bulk.BulkWriteUpsert
 import com.mongodb.client.WriteConcernError
-import com.mongodb.client.result.BulkWriteUpsert
 import org.bson.BsonDocument
 import org.bson.BsonString
 import spock.lang.Specification
